// projects/ngx-ramblers/src/app/models/environment.model.ts
export enum Environment {
  ADMIN_MONGODB_URI = "ADMIN_MONGODB_URI",
  APP_NAME = "APP_NAME",
  AUTH_SECRET = "AUTH_SECRET",
  AWS_ACCESS_KEY_ID = "AWS_ACCESS_KEY_ID",
  AWS_BUCKET = "AWS_BUCKET",
  AWS_REGION = "AWS_REGION",
  AWS_SECRET_ACCESS_KEY = "AWS_SECRET_ACCESS_KEY",
  BASE_URL = "BASE_URL",
  CHROME_VERSION = "CHROME_VERSION",
  CLOUDFLARE_CONFIG = "CLOUDFLARE_CONFIG",
  CMS_PASSWORD = "CMS_PASSWORD",
  CMS_URL = "CMS_URL",
  CMS_USERNAME = "CMS_USERNAME",
  DEBUG = "DEBUG",
  DEBUG_COLORS = "DEBUG_COLORS",
  ENVIRONMENT_SETUP_API_KEY = "ENVIRONMENT_SETUP_API_KEY",
  PLATFORM_ADMIN_ENABLED = "PLATFORM_ADMIN_ENABLED",
  FLICKR_SCRAPE_DUMP_DIR = "FLICKR_SCRAPE_DUMP_DIR",
  FLY_APP_NAME = "FLY_APP_NAME",
  FLY_API_TOKEN = "FLY_API_TOKEN",
  FLY_CONFIG = "FLY_CONFIG",
  FLY_MACHINE_ID = "FLY_MACHINE_ID",
  FLY_METRICS_TOKEN = "FLY_METRICS_TOKEN",
  FLY_ORGANISATION = "FLY_ORGANISATION",
  GH_TOKEN = "GH_TOKEN",
  GITHUB_ACTIONS = "GITHUB_ACTIONS",
  GITHUB_PAT = "GITHUB_PAT",
  GITHUB_TOKEN = "GITHUB_TOKEN",
  IMAGE = "IMAGE",
  MONGODB_URI = "MONGODB_URI",
  NODE_ENV = "NODE_ENV",
  PLAYWRIGHT_HEADLESS = "PLAYWRIGHT_HEADLESS",
  RAMBLERS_FEATURE = "RAMBLERS_FEATURE",
  RAMBLERS_METADATA_FILE = "RAMBLERS_METADATA_FILE",
  RAMBLERS_PASSWORD = "RAMBLERS_PASSWORD",
  INTEGRATION_WORKER_CALLBACK_BASE_URL = "INTEGRATION_WORKER_CALLBACK_BASE_URL",
  INTEGRATION_WORKER_CALLBACK_SECRET = "INTEGRATION_WORKER_CALLBACK_SECRET",
  INTEGRATION_WORKER_APP_NAME = "INTEGRATION_WORKER_APP_NAME",
  INTEGRATION_WORKER_CALLBACK_PROGRESS_PATH = "INTEGRATION_WORKER_CALLBACK_PROGRESS_PATH",
  INTEGRATION_WORKER_CALLBACK_RESULT_PATH = "INTEGRATION_WORKER_CALLBACK_RESULT_PATH",
  INTEGRATION_WORKER_ENCRYPTION_KEY = "INTEGRATION_WORKER_ENCRYPTION_KEY",
  INTEGRATION_WORKER_IMAGE_REPOSITORY = "INTEGRATION_WORKER_IMAGE_REPOSITORY",
  INTEGRATION_WORKER_JOB_ID = "INTEGRATION_WORKER_JOB_ID",
  INTEGRATION_WORKER_MEMORY = "INTEGRATION_WORKER_MEMORY",
  INTEGRATION_WORKER_SCALE_COUNT = "INTEGRATION_WORKER_SCALE_COUNT",
  INTEGRATION_WORKER_SECRETS_JSON = "INTEGRATION_WORKER_SECRETS_JSON",
  INTEGRATION_WORKER_SHARED_SECRET = "INTEGRATION_WORKER_SHARED_SECRET",
  INTEGRATION_WORKER_URL = "INTEGRATION_WORKER_URL",
  RAMBLERS_USERNAME = "RAMBLERS_USERNAME",
  SETUP_AWS_ACCESS_KEY_ID = "SETUP_AWS_ACCESS_KEY_ID",
  SETUP_AWS_REGION = "SETUP_AWS_REGION",
  SETUP_AWS_SECRET_ACCESS_KEY = "SETUP_AWS_SECRET_ACCESS_KEY",
  SKIP_MIGRATIONS_ON_STARTUP = "SKIP_MIGRATIONS_ON_STARTUP",
}